Improved Grass Collision
Grass was colliding with the ground it was placed on, at least if there was a medium sized bump. Fixed by using a Layermask in the collision Raycast function for the grass. Changed Grass Placement layer to "grass" layer and assigned the "grass collision layer" (new) to "default". Now it collides with everything you throw at it but not the ground.